+++ app/src/main/java/khobir/abdul/bareksa_test/domain/perbandingan/model/ChartDataItemModel.kt
@@ -5,5 +5,5 @@ import com.google.gson.annotations.SerializedName
 data class ChartDataItemModel(
     val date: String,
     val value: Float,
-    val growth: Float
+    val growth: Double
 )
